backend/pyproject.toml
[project]
name = "quartiles"
version = "0.1.0"
description = "Quartiles backend API"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard]<1.0.0,>=0.114.2",
    "python-multipart<1.0.0,>=0.0.7",
    "email-validator<3.0.0.0,>=2.1.0.post1",
    "tenacity<9.0.0,>=8.2.3",
    "pydantic>2.0",
    "emails<1.0,>=0.6",
    "jinja2<4.0.0,>=3.1.4",
    "alembic<2.0.0,>=1.12.1",
    "httpx<1.0.0,>=0.25.1",
    "psycopg[binary]<4.0.0,>=3.1.13",
    "sqlmodel<1.0.0,>=0.0.21",
    "pydantic-settings<3.0.0,>=2.2.1",
    "sentry-sdk[fastapi]<2.0.0,>=1.40.6",
    "pyjwt<3.0.0,>=2.8.0",
    "pwdlib[argon2,bcrypt]>=0.3.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ty>=0.0.8",
    "ruff>=0.14.10",
    "prek>=0.2.24,<1.0.0",
    "coverage<8.0.0,>=7.4.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ty.environment]
python = "../.venv"
python-version = "3.13"

[tool.pytest.ini_options]
testpaths = ["app/tests"]

[tool.ruff]
line-length = 120
fix = true
target-version = "py313"
exclude = ["alembic"]

[tool.ruff.lint]
preview = true
select = [
    "FAST", # FastAPI - FastAPI specific rules
    "YTT",  # flake8-2020 - checks for Python 2020 beest practices
    "ANN",  # flake8-annotations - enforces type annotation style
    "S",    # flake8-bandit - security issues
    "BLE",  # flake8-blind-except – flags bare excepts
    "FBT",  # flake8-boolean-trap – potential pitfalls with booleans
    "B",    # flake8-bugbear - common bug patterns
    "A",    # flake8-builtins - misuse of python builtin names
    "C4",   # flake8-comprehensions - best practices in comprehensions
    "DTZ",  # flake8-datetimez – requires timezone-aware datetime objects
    "T10",  # flake8-debugger – debugger statements (e.g. pdb)
    "SIM",  # flake8-simplify
    "EM",   # flake8-errmsg – error message style
    "EXE",  # flake8-executable – executable file checks
    "FA",   # flake8-future-annotations – future import for annotations
    "ISC",  # flake8-implicit-str-concat – warns on implicit string concatenation
    "ICN",  # flake8-import-conventions – enforces conventional import aliases
    "LOG",  # flake8-logging – proper logging usage
    "G",    # flake8-logging-format – logging format string issues
    "INP",  # flake8-no-pep420 – warns against non-PEP420 namespace usage
    "PIE",  # flake8-pie – Python improvement suggestions
    "PYI",  # flake8-pyi – checks for type stub (.pyi) consistency
    "PT",   # flake8-pytest-style – pytest best practices
    "Q",    # flake8-quotes – enforces quote style consistency
    "RSE",  # flake8-raise – proper raise statement usage
    "RET",  # flake8-return – return statement issues
    "SLF",  # flake8-self – flags instance methods that don't use self
    "SLOT", # flake8-slots – suggests use of __slots__ where appropriate
    "SIM",  # flake8-simplify – code simplification hints
    "TID",  # flake8-tidy-imports – enforces specific import styles (e.g., no relative imports)
    "TC",   # flake8-type-checking – proper import of typing in type checks
    "INT",  # flake8-gettext – checks for proper internationalization usage
    "ARG",  # flake8-unused-arguments – flags unused function arguments
    "PTH",  # flake8-use-pathlib – encourages pathlib over os.path
    "TD",   # flake8-todos – flags TODO comments
    "FIX",  # flake8-fixme – flags FIXME comments
    "FLY",  # Flynt - f-string conversion suggestions
    "I",    # isort - import ordering checks
    "C90",  # mccabe - complexity metric for functions
    "NPY",  # NumPy-specific rules - ensures NumPy coding standards
    "PD",   # pandas-vet – checks pandas-specific code practices
    "N",    # PEP8 Naming – naming conventions
    "PERF", # Perflint – performance-related checks
    "E",    # pycodestyle errors (style issues)
    "W",    # pycodestyle warnings (style issues)
    "DOC",  # pydoclint - docstring style
    "D",    # pydocstyle - docstring formatting
    "F",    # pyflakes - detects syntax erros and basic mistakes
    "PGH",  # pygrep-hooks - custom grep hooks for linting
    "UP",   # pyupgrade - upgrades syntax to newer Python versions
    "RUF",  # Ruff-specific rules - additional ruff checks
    "TRY",  # Tryceratops – try/except usage suggestions
]
ignore = [
    "E501", # line-too-long
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
    "B904", # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "S101", "ANN", "BLE", "TRY", "INP001"]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.coverage.run]
branch = true
source = ["app"]

[tool.coverage.report]
skip_empty = true
show_missing = true
sort = "-Cover"

[tool.coverage.html]
show_contexts = true
